requirements.txt
streamlit
pandas
sqlalchemy
psycopg2-binary
plotly
google-generativeai
